build: Remove jsoncpp sources
We have no idea why jsoncpp version 1.7.7 was used as a cutoff,
according to the people packaging for debian, building also works with
jsoncpp 1.7.4 [1]. Also on repology [2], there are barely any distros
with jsoncpp older than 1.7.4.

I very much dislike distributing the entire jsoncpp code with the
project, so I think it's better to remove it.

If jsoncpp is using semantic versioning, even version 1.7.0 should work
and the solution to people having an even older jsoncpp version (1.6.5
was released in 2015) shouldn't be to distribute the source code, but
for them to update their system.

[1]: https://salsa.debian.org/debian/polybar/blob/debian/master/debian/patches/jsoncpp_version.patch
[2]: https://repology.org/project/jsoncpp/versions